refactor: Map 컴포넌트 렌더링 로직 변경
