+++ frontend/public/index.html
@@ -8,7 +8,7 @@
     <meta http-equiv="cache-control" content="max-age=31536000, no-cache" />
     <script
       script
-      src="https://apis.openapi.sk.com/tmap/vectorjs?version=1&appKey=P2MX6F1aaf428AbAyahIl9L8GsIlES04aXS9hgxo"
+      src="https://apis.openapi.sk.com/tmap/jsv2??version=1&appKey=P2MX6F1aaf428AbAyahIl9L8GsIlES04aXS9hgxo"
     ></script>
 
     <title>괜찮을지도</title>
