+++ frontend/src/components/Layout/index.tsx
@@ -1,3 +1,4 @@
+import { useEffect, useRef } from 'react';
 import Map from '../Map';
 import Flex from '../common/Flex';
 import Input from '../common/Input';
@@ -8,7 +9,26 @@ type LayoutProps = {
   children: React.ReactNode;
 };
 
+declare global {
+  interface Window {
+    Tmapv2: any;
+  }
+}
+
 const Layout = ({ children }: LayoutProps) => {
+  const { Tmapv2 } = window;
+  const mapContainer = useRef(null);
+
+  useEffect(() => {
+    const map = new Tmapv2.Map(mapContainer.current, {
+      center: new Tmapv2.LatLng(37.5055, 127.0509),
+    });
+
+    return () => {
+      map.destroy();
+    };
+  }, []);
+
   return (
     <Flex height="100vh" width="100vw">
       <Flex
@@ -29,7 +49,7 @@ const Layout = ({ children }: LayoutProps) => {
           {children}
         </Flex>
       </Flex>
-      <Map />
+      <Map ref={mapContainer} />
     </Flex>
   );
 };
